Baygur: remove redundant UIKit import.
The first part of the exercise includes adding a `LTKit-Prefix.pch`
import, that includes a `UIKit.h` import in it, to `Baygur-Prefix.pch`.
This therefore removes the redundant imports from `main.m`,
`BGRAppDelegate.h` and `BGRViewController.h`.
The import is added to `Baygur-Prefix.pch`, with an option to remove it
as part of the exercise when adding the `LTKit-Prefix.pch` import.